fix for malformed negative latitude/longitude values
The '%' operator does not deal with negative values the way I
intended.  BlueNMEA was never tested in regions west of Greenwich or
on the southern hemisphere, sorry for this bug.  Call Math.abs()
before applying the '%' operator.

In other news, this patch also fixes a rounding error (latitude was
casted to int first, then multiplied).